Split home conversion sections
